+++ NEWS.md
@@ -5,6 +5,7 @@
 * New function `named_vector()`.
 * New function `keep_in_bounds()`.
 * New function `set_col_names()`.
+* New function `distinct_values()`.
 * Additional arguments passed to `Dictionary$get()` and `Dictionary$mget()`
   must now be named.